Prevent trying after fetch without merge via GitPreCheck
Adding a new Git PreCheck to detect if the local working copy has commits behind. When it's the case, the diff generated via "git diff origin/branch" will "undo" the un-merged commits.
As a result, the code 'tried' on the jenkins server will no longer be  the latest copy of the code with the local changes, but instead the exact same code than the local working copy.